requirements.txt
appdirs==1.4.3
arabic-reshaper==2.1.1
autopep8==1.4.4
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
Django==2.2
django-allauth==0.41.0
django-autoslug==1.9.6
django-bootstrap-breadcrumbs==0.9.2
django-countries==6.1
django-crispy-forms==1.9.0
django-extensions==2.2.8
django-import-export==2.0.2
django-modeltranslation==0.15
django-phonenumber-field==4.0.0
django-postgres-dropdb==2019.8.24
django-rosetta==0.9.3
django-translation-flags==1.0.6
et-xmlfile==1.0.1
filelock==3.0.12
future==0.18.2
html5lib==1.1
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.4.0
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
packaging==20.3
pep8==1.7.1
phonenumbers==8.12.1
Pillow==8.3.2
pluggy==0.13.1
polib==1.1.0
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pyparsing==2.4.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-dotenv==0.15.0
python3-openid==3.1.0
pytz==2018.5
PyYAML==5.3.1
reportlab==3.5.55
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.2.4
tablib==1.1.0
toml==0.10.0
tox==3.14.6
unicode==2.7
Unidecode==1.1.1
urllib3==1.25.8
virtualenv==20.0.15
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
django-debug-toolbar==3.2.2
ipython==7.28.0
gunicorn==20.1.0
django-storages==1.12.3
boto3==1.19.10
django-cors-headers==3.10.0